Refactor ActiveAdmin Install
Why:

The AtomicCms generator should act like other generators and provide
useful output to the developer.

This change addresses the need by:

Refactoring install_cms to be more helpful.
